+++ clients/admin-ui/src/features/privacy-experience/Preview.tsx
@@ -166,6 +166,16 @@ const Preview = ({
     );
   }
 
+  if (values.component === ComponentType.PRIVACY_CENTER) {
+    return (
+      <NoPreviewNotice
+        title="Privacy center preview not available"
+        description="There is no preview available for privacy center. You can edit the available settings
+      and languages to the left."
+      />
+    );
+  }
+
   if (!values.privacy_notice_ids?.length) {
     return (
       <NoPreviewNotice